+++ spec/models/sail/setting_spec.rb
@@ -594,4 +594,27 @@ describe Sail::Setting, type: :model do
       expect(subject).to eq(20.0)
     end
   end
+
+  describe ".database_to_file" do
+    subject { described_class.database_to_file }
+
+    before do
+      described_class.create!(name: :setting,
+                              cast_type: :integer,
+                              value: 1,
+                              description: "My setting",
+                              group: "tuners")
+    end
+
+    it "formats database values to yaml for config file" do
+      file = Tempfile.new
+      allow(File).to receive(:open).with("./config/sail.yml", "w").and_yield(file)
+
+      expect(file).to receive(:write).with(
+        "---\nsetting:\n  description: My setting\n  value: '1'\n  group: tuners\n  cast_type: integer\n"
+      )
+
+      subject
+    end
+  end
 end